Added game service with methods for add/edit/remove

+++ SpringData/AutoMappingObjects/Solutions/Exercises/src/main/java/com/example/automappingobjectsexercise/model/entity/Game.java
@@ -26,7 +26,7 @@ public class Game {
     @NonNull
     private String imageThumbnail;
     @NonNull
-    private Integer size;
+    private Double size;
     @NonNull
     @Column(precision = 9, scale = 2)
     private BigDecimal price;
